+++ packages/client/utils/isViewerTyping.js
@@ -1,5 +1,0 @@
-const isViewerTyping = () => {
-  return document.activeElement.contentEditable === 'true'
-}
-
-export default isViewerTyping