+++ packages/server/postgres/queries/updateUser.ts
@@ -2,12 +2,16 @@ import {updateUserQuery, IUpdateUserQueryParams} from './generated/updateUserQue
 import getPg from '../getPg'
 import catchAndLog from '../utils/catchAndLog'
 import AuthIdentity from '../../database/types/AuthIdentity'
+import {OptionalExceptFor} from '../../utils/TypeUtil'
 
 interface UpdateUserQueryParams extends Omit<IUpdateUserQueryParams, 'identities'> {
   identities: AuthIdentity[]
 }
 
-const updateUser = async (update: Partial<UpdateUserQueryParams>, userIds: string | string[]) => {
+const updateUser = async (
+  update: OptionalExceptFor<UpdateUserQueryParams, 'updatedAt'>,
+  userIds: string | string[]
+) => {
   userIds = typeof userIds === 'string' ? [userIds] : userIds
   return await catchAndLog(() =>
     updateUserQuery.run(